Merge adding of Electrical Overview document
Add business overview as a git submodule

Added the business electrical overview repository as a submodule. This
is temporary as the next commit will remove the submodule reference.
This is for us to remove the overview as a separate repository.

Remove submodule misc/busi_elec_overview

Initial commit

Set up project files, imported all images to be used, made a basic pdf
document, added a readme.md file and a .gitignore

Add Section on PCB design work

Completed the introduction and the section on the PCB team. This
includes descriptions of the driver controls board, CCS, lights board
and audio board.

Add additional images to project

Add a image of the lights turned on and a image of the driver control
board to the images directory

Fix readme file

The header at the top of the readme file had an issue caused by no space
between the # and the rest of the header line.

Add section on systems team

Added a section on the electrical systems team, including the charge
system, motors, motor controllers, and lights.

Add systems schematic to systems overview

Added the systems schematics image from google drive to the systems
overview section of the document

Add battery section

Added a section describing the functionality and safety features of the
battery in the solar car

Add section on arrays

Added the section discussing the solar arrays to the document

Add information on MPPTs

Added a description of the mppts, a few pictures for the mppts.

Remove overview.pdf from repo

Removed teh overview.pdf file from the repository to prevent binary file
merge conflicts

Fix MPPT description to be more technically accurate

Fix formatting issues in battery sections

Update title page

The title page was updated to match the template being used for
documentation in the electrical documentation repository. The logo was
changed in this the wide logo with team name.

Remove helper script to merge submodule

A helper script used to merge the submodule for the business overview
was left in the repository. It is now removed.

Create LaTeX Style guide

Create the beginnings of a style guide for the solar car team to use for
our LaTeX code.

Change style guide to adoc file

Change the style guide to be a adoc file to match with git standards